Add AZSplitView
Fix some other small bugs as well, and implement 'hidden' as fallout from implementing AZSplitView. Also only check for mouse clicks in the visibleRect now.

+++ Azoth/Classes/Drawing/AZRenderer.h
@@ -114,7 +114,7 @@ struct SDL_FPoint;
 \*****************************************************************************/
 - (NSSize) presentationSize;
 - (int) presentationMode;
-- (void) setPresentationSize:(NSSize)size mode:(int)mode;
+- (void) setPresentationSize:(NSSize)size mode:(NSInteger)mode;
 
 
 /*****************************************************************************\
